+++ filenames.gni
@@ -433,10 +433,8 @@ filenames = {
     "shell/common/api/atom_api_native_image_mac.mm",
     "shell/common/api/atom_api_shell.cc",
     "shell/common/api/atom_api_v8_util.cc",
-    "shell/common/api/constructor.h",
     "shell/common/api/electron_bindings.cc",
     "shell/common/api/electron_bindings.h",
-    "shell/common/api/constructor.h",
     "shell/common/api/features.cc",
     "shell/common/application_info.cc",
     "shell/common/application_info.h",
@@ -496,6 +494,7 @@ filenames = {
     "shell/common/gin_helper/arguments.h",
     "shell/common/gin_helper/callback.cc",
     "shell/common/gin_helper/callback.h",
+    "shell/common/gin_helper/constructor.h",
     "shell/common/gin_helper/destroyable.cc",
     "shell/common/gin_helper/destroyable.h",
     "shell/common/gin_helper/dictionary.h",
