+++ shell/browser/api/views/atom_api_button.cc
@@ -4,7 +4,7 @@
 
 #include "shell/browser/api/views/atom_api_button.h"
 
-#include "shell/common/api/constructor.h"
+#include "shell/common/gin_helper/constructor.h"
 #include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_includes.h"
 
@@ -50,7 +50,7 @@ void Initialize(v8::Local<v8::Object> exports,
                 void* priv) {
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
-  dict.Set("Button", mate::CreateConstructor<Button>(
+  dict.Set("Button", gin_helper::CreateConstructor<Button>(
                          isolate, base::BindRepeating(&Button::New)));
 }
