+++ shell/browser/api/views/atom_api_label_button.cc
@@ -5,7 +5,7 @@
 #include "shell/browser/api/views/atom_api_label_button.h"
 
 #include "base/strings/utf_string_conversions.h"
-#include "shell/common/api/constructor.h"
+#include "shell/common/gin_helper/constructor.h"
 #include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_includes.h"
 
@@ -70,7 +70,7 @@ void Initialize(v8::Local<v8::Object> exports,
                 void* priv) {
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
-  dict.Set("LabelButton", mate::CreateConstructor<LabelButton>(
+  dict.Set("LabelButton", gin_helper::CreateConstructor<LabelButton>(
                               isolate, base::BindRepeating(&LabelButton::New)));
 }
