+++ shell/browser/api/views/atom_api_layout_manager.cc
@@ -4,7 +4,7 @@
 
 #include "shell/browser/api/views/atom_api_layout_manager.h"
 
-#include "shell/common/api/constructor.h"
+#include "shell/common/gin_helper/constructor.h"
 #include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_includes.h"
 
@@ -54,7 +54,7 @@ void Initialize(v8::Local<v8::Object> exports,
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
   dict.Set("LayoutManager",
-           mate::CreateConstructor<LayoutManager>(
+           gin_helper::CreateConstructor<LayoutManager>(
                isolate, base::BindRepeating(&LayoutManager::New)));
 }
