+++ shell/browser/api/views/atom_api_resize_area.cc
@@ -4,7 +4,7 @@
 
 #include "shell/browser/api/views/atom_api_resize_area.h"
 
-#include "shell/common/api/constructor.h"
+#include "shell/common/gin_helper/constructor.h"
 #include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_includes.h"
 
@@ -50,7 +50,7 @@ void Initialize(v8::Local<v8::Object> exports,
                 void* priv) {
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
-  dict.Set("ResizeArea", mate::CreateConstructor<ResizeArea>(
+  dict.Set("ResizeArea", gin_helper::CreateConstructor<ResizeArea>(
                              isolate, base::BindRepeating(&ResizeArea::New)));
 }
