+++ shell/browser/api/views/atom_api_text_field.cc
@@ -4,7 +4,7 @@
 
 #include "shell/browser/api/views/atom_api_text_field.h"
 
-#include "shell/common/api/constructor.h"
+#include "shell/common/gin_helper/constructor.h"
 #include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_includes.h"
 
@@ -57,7 +57,7 @@ void Initialize(v8::Local<v8::Object> exports,
                 void* priv) {
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
-  dict.Set("TextField", mate::CreateConstructor<TextField>(
+  dict.Set("TextField", gin_helper::CreateConstructor<TextField>(
                             isolate, base::BindRepeating(&TextField::New)));
 }
