+++ shell/common/gin_helper/constructor.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef SHELL_COMMON_API_CONSTRUCTOR_H_
-#define SHELL_COMMON_API_CONSTRUCTOR_H_
+#ifndef SHELL_COMMON_GIN_HELPER_CONSTRUCTOR_H_
+#define SHELL_COMMON_GIN_HELPER_CONSTRUCTOR_H_
 
-#include "native_mate/constructor.h"
+#include "shell/common/gin_helper/function_template.h"
 
-namespace mate {
+namespace gin_helper {
 
 // Create a FunctionTemplate that can be "new"ed in JavaScript.
 // It is user's responsibility to ensure this function is called for one type
@@ -22,13 +22,13 @@ v8::Local<v8::Function> CreateConstructor(
   CHECK(!called) << "CreateConstructor can only be called for one type once";
   called = true;
 #endif
-  v8::Local<v8::FunctionTemplate> templ = CreateFunctionTemplate(
+  v8::Local<v8::FunctionTemplate> templ = mate::CreateFunctionTemplate(
       isolate, base::BindRepeating(&mate::internal::InvokeNew<Sig>, func));
   templ->InstanceTemplate()->SetInternalFieldCount(1);
   T::BuildPrototype(isolate, templ);
   return templ->GetFunction(isolate->GetCurrentContext()).ToLocalChecked();
 }
 
-}  // namespace mate
+}  // namespace gin_helper
 
-#endif  // SHELL_COMMON_API_CONSTRUCTOR_H_
+#endif  // SHELL_COMMON_GIN_HELPER_CONSTRUCTOR_H_
